back/accounts/views.py
from django.contrib.auth import get_user_model
from rest_framework.generics import CreateAPIView
from rest_framework.permissions import AllowAny

from accounts.serializers import SignupSerializer


class SignupView(CreateAPIView):
    model = get_user_model()
    serializer_class = SignupSerializer

    #회원가입은 인증이 필요없다.
    permission_classes = [
        AllowAny,
    ]