back/requirements/common.txt
django
pillow
djangorestframework
mysqlclient
djangorestframework-jwt
django-cors-headers